fix/쿼리dsl 구조를 동작하게끔 다시 수정
