+++ src/main/java/com/nhnacademy/taskapi/coupon/repository/coupons/CouponBoxRepository.java
@@ -8,7 +8,5 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CouponBoxRepository extends JpaRepository<IssuedCoupon, Long> {
-    Page<IssuedCoupon> findByMemberAndCoupon_CouponStatus_Name(Member member, String couponCouponStatusName, Pageable pageable);
 
-    Page<IssuedCoupon> getValidIssuedCoupon(Long memberId, CouponStatus couponStatus, Pageable pageable);
 }
