+++ src/main/java/com/nhnacademy/taskapi/member/service/impl/MemberServiceImpl.java
@@ -117,7 +117,7 @@ public class MemberServiceImpl implements MemberService {
         try {
 
             Member result = memberRepository.save(member);
-            pointService.registerMemberPoints(result);
+//            pointService.registerMemberPoints(result);
             return MemberResponseDto.from(result);
 
         }catch(DataIntegrityViolationException e) {
